requirements.txt
fastapi
uvicorn
requests
PyMuPDF
lancedb
sentence-transformers
groq
dotenv
python-multipart
torch
easyocr
pdfplumber
pillow
git+https://github.com/brandonstarxel/chunking_evaluation.git
